# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/billing/plans": {
      "get": {
        "operationId": "BillingController_listPlans_v1",
        "parameters": [
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingPlansResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Billing"
        ]
      }
    },
...
    "/api/v1/billing/subscription-setup": {
      "get": {
        "operationId": "BillingController_getSubscriptionSetup_v1",
        "parameters": [
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingSubscriptionSetupResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Billing"
        ]
      }
    },
...
    "/api/v1/billing/subscriptions": {
      "post": {
        "operationId": "BillingController_subscribeCompany_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscribeCompanyResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/subscriptions/{subscriptionId}/attach-payment-method": {
      "post": {
        "operationId": "BillingController_attachPaymentMethod_v1",
        "parameters": [
          {
            "name": "subscriptionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachCompanySubscriptionPaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachPaymentMethodResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/subscriptions/{subscriptionId}/cancel": {
      "post": {
        "operationId": "BillingController_cancelSubscription_v1",
        "parameters": [
          {
            "name": "subscriptionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelCompanySubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelSubscriptionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Billing"
        ]
      }
    },
...
    "/api/v1/company/units/{unitId}/roles": {
      "get": {
        "operationId": "BranchRolesController_findAll_v1",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "BranchRoles"
        ]
      },
      "post": {
        "operationId": "BranchRolesController_create_v1",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "BranchRoles"
        ]
      }
    },
    "/api/v1/company/units/{unitId}/roles/permission-catalog": {
      "get": {
        "operationId": "BranchRolesController_getPermissionCatalog_v1",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "BranchRoles"
        ]
      }
    },
    "/api/v1/company/units/{unitId}/roles/{roleId}": {
      "get": {
        "operationId": "BranchRolesController_findOne_v1",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "BranchRoles"
        ]
      },
      "patch": {
        "operationId": "BranchRolesController_update_v1",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "BranchRoles"
        ]
      }
    },
    "/api/v1/company/units/{unitId}/roles/{roleId}/status": {
      "patch": {
        "operationId": "BranchRolesController_updateStatus_v1",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchRoleStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "BranchRoles"
        ]
      }
    },
...
    "/api/v1/companies/{companyId}/user-sidebar/customization": {
      "get": {
        "operationId": "UserSidebarController_getCustomization_v1",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "branchUnitId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserSidebar"
        ]
      },
      "put": {
        "operationId": "UserSidebarController_save_v1",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "branchUnitId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveUserSidebarDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserSidebar"
        ]
      }
    },
    "/api/v1/companies/{companyId}/user-sidebar/reset": {
      "post": {
        "operationId": "UserSidebarController_reset_v1",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "branchUnitId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "applyScope",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "UserSidebar"
        ]
      }
    },
...
  },
  "info": {
    "title": "GR8Books Neo API",
    "description": "Versioned REST API for GR8Books Neo frontend clients.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BillingPlanPricingResponseDto": {
        "type": "object",
        "properties": {
          "monthly": {
            "$ref": "#/components/schemas/BillingPlanPriceResponseDto"
          },
          "yearly": {
            "$ref": "#/components/schemas/BillingPlanPriceResponseDto"
          }
        },
        "required": [
          "monthly",
          "yearly"
        ]
      },
      "BillingPlanDetailedPriceResponseDto": {
        "type": "object",
        "properties": {
          "amountInCents": {
            "type": "number",
            "nullable": true
          },
          "compareAtInCents": {
            "type": "number",
            "nullable": true
          },
          "isRemoteReady": {
            "type": "boolean"
          },
          "id": {
            "type": "number"
          },
          "billingCycle": {
            "enum": [
              "MONTHLY",
              "QUARTERLY",
              "YEARLY"
            ],
            "type": "string"
          },
          "intervalCount": {
            "type": "number"
          },
          "intervalUnit": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "amountInCents",
          "compareAtInCents",
          "isRemoteReady",
          "id",
          "billingCycle",
          "intervalCount",
          "intervalUnit",
          "isActive"
        ]
      },
      "BillingPlanUsageRuleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "metric": {
            "type": "string"
          },
          "freeCount": {
            "type": "number"
          },
          "unitPriceInCents": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "metric",
          "freeCount",
          "unitPriceInCents"
        ]
      },
...
      "BillingPlanResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "trialDays": {
            "type": "number"
          },
          "pricing": {
            "$ref": "#/components/schemas/BillingPlanPricingResponseDto"
          },
          "prices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillingPlanDetailedPriceResponseDto"
            }
          },
          "usageRules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillingPlanUsageRuleResponseDto"
            }
          },
          "discountTiers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillingPlanDiscountTierResponseDto"
            }
          },
          "moduleKeys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillingPlanModuleResponseDto"
            }
          }
        },
        "required": [
          "code",
          "name",
          "description",
          "currency",
          "scope",
          "trialDays",
          "pricing",
          "prices",
          "usageRules",
          "discountTiers",
          "moduleKeys",
          "modules"
        ]
      },
      "BillingPlansResponseDto": {
        "type": "object",
        "properties": {
          "plans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillingPlanResponseDto"
            }
          }
        },
        "required": [
          "plans"
        ]
      },
...
      "CompanySubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "status": {
            "enum": [
              "INCOMPLETE",
              "TRIALING",
              "ACTIVE",
              "PAST_DUE",
              "UNPAID",
              "INCOMPLETE_CANCELED",
              "EXPIRED",
              "CANCELED"
            ],
            "type": "string"
          },
          "billingCycle": {
            "enum": [
              "MONTHLY",
              "QUARTERLY",
              "YEARLY"
            ],
            "type": "string"
          },
          "billingProvider": {
            "type": "string",
            "enum": [
              "PAYMONGO"
            ]
          },
          "startsAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "trialEndsAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "currentPeriodStartAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "nextBillingAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "endsAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "canceledAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "cancelAtPeriodEnd": {
            "type": "boolean"
          },
          "failureCode": {
            "type": "string",
            "nullable": true
          },
          "failureMessage": {
            "type": "string",
            "nullable": true
          },
          "providerReferences": {
            "$ref": "#/components/schemas/CompanySubscriptionProviderReferencesResponseDto"
          },
          "plan": {
            "$ref": "#/components/schemas/CompanySubscriptionPlanResponseDto"
          },
          "billingCustomer": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/BillingCustomerResponseDto"
              }
            ]
          },
          "paymentMethods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillingPaymentMethodResponseDto"
            }
          },
          "invoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillingInvoiceResponseDto"
            }
          }
        },
        "required": [
          "id",
          "status",
          "billingCycle",
          "billingProvider",
          "startsAt",
          "trialEndsAt",
          "currentPeriodStartAt",
          "nextBillingAt",
          "endsAt",
          "canceledAt",
          "cancelAtPeriodEnd",
          "failureCode",
          "failureMessage",
          "providerReferences",
          "plan",
          "billingCustomer",
          "paymentMethods",
          "invoices"
        ]
      },
      "BillingSubscriptionSetupResponseDto": {
        "type": "object",
        "properties": {
          "plans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillingPlanResponseDto"
            }
          },
          "subscription": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CompanySubscriptionResponseDto"
              }
            ]
          }
        },
        "required": [
          "plans",
          "subscription"
        ]
      },
...
      "SubscribeCompanyDto": {
        "type": "object",
        "properties": {
          "planCode": {
            "type": "string"
          },
          "billingCycle": {
            "enum": [
              "MONTHLY",
              "QUARTERLY",
              "YEARLY"
            ],
            "type": "string"
          }
        },
        "required": [
          "planCode",
          "billingCycle"
        ]
      },
      "SubscribeCompanyResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "subscription": {
            "$ref": "#/components/schemas/CompanySubscriptionResponseDto"
          }
        },
        "required": [
          "message",
          "subscription"
        ]
      },
      "AttachCompanySubscriptionPaymentMethodDto": {
        "type": "object",
        "properties": {
          "paymentMethodId": {
            "type": "string"
          }
        },
        "required": [
          "paymentMethodId"
        ]
      },
      "PaymentIntentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "redirectUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "status",
          "redirectUrl"
        ]
      },
      "AttachPaymentMethodResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "subscription": {
            "$ref": "#/components/schemas/CompanySubscriptionResponseDto"
          },
          "paymentIntent": {
            "$ref": "#/components/schemas/PaymentIntentResponseDto"
          },
          "pendingProviderActivation": {
            "type": "boolean"
          }
        },
        "required": [
          "message",
          "subscription",
          "paymentIntent"
        ]
      },
      "CancelCompanySubscriptionDto": {
        "type": "object",
        "properties": {
          "cancelAtPeriodEnd": {
            "type": "boolean"
          }
        }
      },
      "CancelSubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "subscription": {
            "$ref": "#/components/schemas/CompanySubscriptionResponseDto"
          }
        },
        "required": [
          "message",
          "subscription"
        ]
      },
...
      "BranchRolePermissionDto": {
        "type": "object",
        "properties": {
          "moduleCode": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "moduleName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160
          },
          "permissionCode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "permissionName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160
          },
          "actions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "view",
                "create",
                "update",
                "cancel",
                "uncancel",
                "export"
              ]
            }
          },
          "canView": {
            "type": "boolean"
          },
          "canCreate": {
            "type": "boolean"
          },
          "canUpdate": {
            "type": "boolean"
          },
          "canCancel": {
            "type": "boolean"
          },
          "canUncancel": {
            "type": "boolean"
          },
          "canExport": {
            "type": "boolean"
          }
        },
        "required": [
          "moduleCode",
          "moduleName",
          "permissionCode",
          "permissionName"
        ]
      },
      "CreateBranchRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000
          },
          "permissions": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BranchRolePermissionDto"
            }
          }
        },
        "required": [
          "name",
          "permissions"
        ]
      },
      "UpdateBranchRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000
          },
          "permissions": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BranchRolePermissionDto"
            }
          }
        },
        "required": [
          "name",
          "permissions"
        ]
      },
      "UpdateBranchRoleStatusDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "isActive"
        ]
      },
...
      "UserSidebarTreeItemDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "maxLength": 120
          },
          "label": {
            "type": "string",
            "maxLength": 160
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "itemType": {
            "type": "string",
            "enum": [
              "SECTION",
              "CONTAINER",
              "LINK"
            ]
          },
          "moduleId": {
            "type": "number",
            "minimum": 1
          },
          "iconName": {
            "type": "string"
          },
          "isHidden": {
            "type": "boolean"
          },
          "isPinned": {
            "type": "boolean"
          },
          "isCollapsed": {
            "type": "boolean"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSidebarTreeItemDto"
            }
          }
        },
        "required": [
          "key",
          "label",
          "itemType",
          "children"
        ]
      },
      "SaveUserSidebarDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 0
          },
          "applyScope": {
            "enum": [
              "CURRENT_BRANCH",
              "ALL_BRANCHES"
            ],
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSidebarTreeItemDto"
            }
          }
        },
        "required": [
          "version",
          "items"
        ]
      },
...
    }
  }
}